# api/src/resources/ocapi_openapi.yaml
openapi: 3.0.2
info:
  description: This is a calculation interface from Mercedes-Benz Mobility AG.
  version: 3.17.0
  title: OCAPI
  termsOfService: https://git.i.mercedes-benz.com/mbm-apis/OneCalculationAPI
  contact:
    email: ocapi@mercedes-benz.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: >-
      https://{environment}.api.mercedes-benz.com/v3
    variables:
      environment:
        default: ocapi  # Production server
        enum:
          - ocapi       # Production server
          - dev.ocapi   # Development server
          - test.ocapi  # Testing server
paths:
  /{market}/calculations:
    post:
      tags:
        - calculations
      summary: Calculate a new financing quote.
      description: Calculate a new financing quote and return all that is needed to render
        the input fields as well as the result.
      operationId: calculations
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: market
          description: Enabled markets - ISO 3166-1 alpha-2 country code of a market
          required: true
          schema:
            type: string
            default: de
            enum:         # Enabled markets - ISO 3166-1 alpha-2 country code of a market
              - at
              - au
              - be
              - ch
              - cz
              - de
              - dk
              - es
              - fr
              - gb
              - hu
              - in
              - it
              - jp
              - kr
              - lu
              - mx
              - my
              - nl
              - nz
              - pl
              - pt
              - ro
              - se
              - sg
              - sk
              - th
              - tr
              - tw
              - za
        - in: header
          name: Accept-Language
          description: 2 letter code ISO 639-1 for language followed by a hyphen and ISO 3166-1 ALPHA-2 of the country
          example: de-DE
          schema:
            type: string
          required: false
        - in: header
          name: User-Agent
          description: currently not used in V3, default values can be used
          example: Salestablet/v1.2
          schema:
            type: string
          required: false
        - in: header
          name: X-OCAPI-User-Interaction
          description: The rendering UI can send information on whether the customer has seen or changed the calculation. Used for measuring marketing acceptance.
          example: calculator-viewed or output-viewed or short-summary-viewed or calculation-changed
          schema:
            type: string
          required: false
        - in: header
          name: X-Request-ID
          description: Unique identifier of the request. Makes it easier to trace behavior and trouble shooting.
          example: 77e1c83b-7bb0-437b-bc50-a7a58e5660ac
          schema:
            type: string
            format: uuid
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculationModel"
            examples:
              de:
                $ref: '#/components/examples/de'
              pt:
                $ref: '#/components/examples/pt'
        description: Calculation context that needs to be used.
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalcResponse"
        "400":
          description: invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                "40010":
                  $ref: '#/components/examples/40010'
                "40030":
                  $ref: '#/components/examples/40030'
                "40031":
                  $ref: '#/components/examples/40031'
                "40040":
                  $ref: '#/components/examples/40040'
                "40050":
                  $ref: '#/components/examples/40050'
                "40060":
                  $ref: '#/components/examples/40060'
                "40061":
                  $ref: '#/components/examples/40061'
                "40062":
                  $ref: '#/components/examples/40062'
                "40063":
                  $ref: '#/components/examples/40063'
                "40070":
                  $ref: '#/components/examples/40070'
                "40071":
                  $ref: '#/components/examples/40071'
                "40072":
                  $ref: '#/components/examples/40072'
        "403":
          description: forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                "40310":
                  $ref: '#/components/examples/40310'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                "50060":
                  $ref: '#/components/examples/50060'
                "50070":
                  $ref: '#/components/examples/50070' 
                "50080":
                  $ref: '#/components/examples/50080'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-Key is required for consumers only (not for providers)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponseBody"
...
    CalculationModel:
      type: object
      description: "CalculationModel"
      required:
        - vehicle
        - input
      properties:
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        customer:
          $ref: "#/components/schemas/Customer"
        input:
          type: array
          items:
            $ref: "#/components/schemas/FinancialParameters"
        opaque:
          type: string
          description: empty string on initial request, otherwise must be the value of the opaque value returned in the last reponse
    Customer:
      type: object
      properties:
        type:
          type: string
          description: type of the customer
          enum:
            - private
            - business
        segmentId:
          type: string
          description: customer segment ID
        fleetAccountNumber:
          type: string
          description: number to group customer in a customer segment campaign or helps to identify a specific campaign for a customer
        productType:
          type: string
          description: productType to be displayed as a default (ranked first)
          enum:
            - leasing
            - financing
        subProductType:
          type: string
          description: subProductType to be displayed as a default (ranked first)
          enum:
            - financing_standard
            - financing_balloon
            - financing_option
            - leasing_operating
            - leasing_operating_options
            - leasing_finance
            - leasing_finance_options
            - leasing_finance_balloon
        regionCode:
          type: string
          description: region code of the customer, e.g. zip code
        financialAttributes:
          type: array
          description: desired financial attributes that can be taken into account in the calculation
          items:
            type: object
            properties:
              id:
                type: string
                enum:
                  - downpayment
                  - targetMonthlyInstallment
                  - contractDuration
                  - annualMileage
                  - initialPaymentMultiple
              currency:
                type: string
                description: currency of the financial attribute ( deprecated, use unit instead )
                example: EUR
              unit:
                type: string
                description: unit of that specific item ( i.e. km, mi or miles, month or any valid currency )   
                example: month
              rawValue:
                type: number
                description: raw value of the financial attribute
                example: 10000.00
        tradeIn:
          $ref: "#/components/schemas/TradeIn"
...
    Vehicle:
      type: object
      description: "Vehicle"
      properties:
        technicalIdReference:
          type: string
          description: some reference to an (existing i.e. already built) vehicle; depending on sales region and market this might be VDS CommissionNo, VX VehicleId or VerSO-ConfigID.
        vdzDeal:
          type: boolean
          description: transaction is a Model D VdZ transaction (true) or non-VdZ transaction (false)
        vdzType:
          type: string
          description: flag used to identify different steps during migration ( temporary field, to be removed after migration )
          enum:
            - nonVdZ
            - SGL
            - VdZBlueprint
        dealId:
          type: string
          description: some reference to the conditions of prices and subsidies, e.g. UID of Pricing Engine
        fin:
          type: string
          description: vehicle FIN or VIN, limited use within EU member states, see OCAPI guide for more details
        alternativeConfiguration:
          type: array
          description: possible equipments to be chosen in configurator to calculate partial instalments (instalments on equipment level)
          items:
            $ref: "#/components/schemas/Equipment"
        vehicleConfiguration:
          $ref: "#/components/schemas/VehicleConfiguration"
        condition:
          $ref: "#/components/schemas/Condition"
        technicalData:
          type: array
          items:
            $ref: "#/components/schemas/TechnicalData"
        name:
          type: string
          description: name of the model
          example: C 160 Limousine
        vatReclaimable:
          type: boolean
          description: flag that indicates whether VAT can be claimed ( true ) or not. Applicable for used cars.
        prices:
          $ref: "#/components/schemas/PriceInfoContainer"
        additionalValues:
          $ref: "#/components/schemas/AdditionalValuesContainer"
        financialCode:
          type: string
          description: code (if available) for the current configuration
        usage:
          type: string
          description: indicates the usage type of the vehicle (taxi, driving instructor vehicle,...)
          example: taxi
        warranty:
          type: string
          description: type of warranty, e.g. CPO
          enum:
            - CPO
        quantity:
          type: number
          description: number of same vehicles the calculation is requested for
        additionalVehicleAttributes:
          type: array
          description: all kind of additional attributes to the vehicle
          items:
            type: object
            properties:
              id:
                type: string
                enum:
                  - drehscheibe  #special vehicle pool in market de
                  - jungeSterne
                  - preGrounding
                  - postGrounding
        BAFA:
          type: object
          description: Properties for BAFA document to receive environmental bonus in Germany
          properties:
            specialEquipmentsNet:
              type: number
              description: total sum of all equipments
            onTheRoadPriceNet:
              type: number
              description: acquisition costs
            extraCostsNet:
              type: number
              description: total amount of the extended scope of delivery
            deliveryChargeNet:
              type: number
              description: transfer costs
        vehicleOrderDate:
          type: string
          description: date of vehicle order in ISO-8601 format w/o time portion, i.e. YYYY-MM-DD
          example: 2024-07-23
        taxCase:
          type: string
          enum:
            - standard
            - nato
            - gClassInGraz
            - depot
          description: tax cases which apply for the whole deal/vehicle (relevant for accounting)
          example: "nato"
        proposalType:
          type: string
          description: type of proposal. Indicates whether the request is an initial proposal ("initial"), a proposal change ("change") or a follow-up financing ("followUp").
          default: "initial"
          enum:
            - initial
            - change
            - followUp
        initialProposalDate:
          type: string
          description: date of initial proposal in ISO-8601 format w/o time portion, i.e. YYYY-MM-DD. The date shows the day on which the customer creates the initial proposal, since the date determines the residual value which will be used during calculations of future contract changes.
          example: "2024-09-27"
        proposalVersion:
          type: number
          description: Common identifier for calculation/order version. Needed to handle order amendment in stationary sales, created by vehicle POS system
          example: 1
...
    FinancialParameters:
      type: object
      required:
        - id
        - value
      properties:
        id:
          type: string
          example: term
        value:
          type: string
          example: "12"
    CalcResponse:
      type: object
      description: "CalcResponse"
      properties:
        output:
          $ref: "#/components/schemas/Output"
        input:
          $ref: "#/components/schemas/InputContainer"
        tables:
          $ref: "#/components/schemas/TableContainer"
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
        opaque:
          type: string
          description: if this is returned in a response, the subsequent request has to add this property
...
    InputContainer:
      type: object
      description: "InputContainer"
      properties:
        id:
          type: string
          description: unique id of the container.
          enum:
            - input
            - service
        label:
          type: string
          description: label of the container (e.g. to be displayed in UI-frontend)
        items:
          type: array
          items:
            $ref: "#/components/schemas/InputField"
        childContainers:
          type: array
          items:
            $ref: "#/components/schemas/InputContainer"
    TableContainer:
      type: array
      description: contains all tables
      items:
        $ref: "#/components/schemas/table"
...
  examples:
    de:
      description: "example DE"
      summary: An example for de market
      value:
        input: []
        vehicle:
          vehicleConfiguration:
            baumuster: '2053401'
            nst: '999'
            modelYear: '809'
            salesDescription: 'C 180 Coupé'
            equipments: []
          name: 'C 180 Coupé'
          prices:
          - id: grossListPrice
            currency: EUR
            rawValue: 37549.2
            charges: []
          - id: baseListPrice
            currency: EUR
            rawValue: 37549.2
            charges: []
    pt:
      description: "example pt"
      summary: An example for pt market
      value:
        input: []
        vehicle:
          vehicleConfiguration:
            baumuster: '2053401'
            nst: '999'
            modelYear: '809'
            salesDescription: 'C 180 Coupé'
            equipments: []
          name: 'C 180 Coupé'
          prices:
          - id: grossListPrice
            currency: EUR
            rawValue: 37549.2
            charges: []
          - id: baseListPrice
            currency: EUR
            rawValue: 37549.2
            charges: []
    "40010":
      description: "An example for error code 40010"
      summary: An example for error code 40010
      value:
        error:
            status: 400
            code: 40010
            message: 'invalid request: request parsing error'
    "40030":
      description: "An example for error code 40030"
      summary: An example for error code 40030
      value:
        error:
            status: 400
            code: 40030
            message: 'invalid request: missing or unknown market'
    "40031":
      description: "An example for error code 40031"
      summary: An example for error code 40031
      value:
        error:
            status: 400
            code: 40031
            message: 'invalid request: no data found for market:%iso%, module:%stock_car%, language:%iso%'
    "40040":
      description: "An example for error code 40040"
      summary: An example for error code 40040
      value:
        error:
            status: 400
            code: 40040
            message: 'invalid request: empty or unknown caller'
    "40050":
      description: "An example for error code 40050"
      summary: An example for error code 40050
      value:
        error:
            status: 400
            code: 40050
            message: 'invalid request: empty locale'
    "40060":
      description: "An example for error code 40060"
      summary: An example for error code 40060
      value:
        error:
            status: 400
            code: 40060
            message: 'invalid request: vehicle is missing'
    "40061":
      description: "An example for error code 40061"
      summary: An example for error code 40061
      value:
        error:
            status: 400
            code: 40061
            message: 'invalid request: error in vehicle identification'
    "40062":
      description: "An example for error code 40062"
      summary: An example for error code 40062
      value:
        error:
            status: 400
            code: 40062
            message: 'invalid request: no funding product configured for this vehicle'
    "40063":
      description: "An example for error code 40063"
      summary: An example for error code 40063
      value:
        error:
            status: 400
            code: 40063
            message: 'invalid request: no calculation provided for this retailer'
    "40070":
      description: "An example for error code 40070"
      summary: An example for error code 40070
      value:
        error:
            status: 400
            code: 40070
            message: 'invalid request:  js schema validator message'
    "40071":
      description: "An example for error code 40071"
      summary: An example for error code 40071
      value:
        error:
            status: 400
            code: 40071
            message: 'invalid request: plausability check failed (field names)'
    "40072":
      description: "An example for error code 40072"
      summary: An example for error code 40072
      value:
        error:
            status: 400
            code: 40072
            message: 'invalid request: validation check failed, limits are reached (field name and limit info)'
    "40310":
      description: "An example for error code 40310"
      summary: An example for error code 40310
      value:
        error:
            status: 403
            code: 40010
            message: 'forbidden'
    "50060":
      description: "An example for error code 50060"
      summary: An example for error code 50060
      value:
        error:
            status: 500
            code: 50060
            message: 'server error: unclassified calculation error'
    "50070":
      description: "An example for error code 50070"
      summary: An example for error code 50070
      value:
        error:
            status: 500
            code: 50070
            message: 'server error: fatal error during request processing + message'
    "50080":
      description: "An example for error code 50080"
      summary: An example for error code 50080
      value:
        error:
            status: 500
            code: 50080
            message: 'server error: action with dependent ressource failed'